Refactor features and add new services in order handling

+++ app/Services/Verification/VerifyEmail.php
@@ -2,11 +2,13 @@
 
 namespace App\Services\Verification;
 
+use App\Models\Admin;
 use Carbon\Carbon;
 use Illuminate\Support\Str;
 use App\Models\Verification;
 use App\Traits\ResponseTrait;
 use Illuminate\Http\JsonResponse;
+use Illuminate\Support\Facades\Auth;
 use App\Interfaces\Authentication\VerificationInterface;
 
 class VerifyEmail implements VerificationInterface
@@ -16,12 +18,12 @@ class VerifyEmail implements VerificationInterface
     private $auth;
     public function __construct()
     {
-        $this->auth = auth()->user();
+        $this->auth = Auth::user();
     }
     public function sendNotification(): JsonResponse
     {
 
-        if(!$this->auth->hasVerifiedEmail())
+        if($this->auth instanceof Admin &&  !$this->auth->hasVerifiedEmail())
         {
             $this->auth->sendEmailVerificationWithToken();
 
@@ -32,25 +34,29 @@ class VerifyEmail implements VerificationInterface
     }
     public function verify($token): JsonResponse
     {
-        $verification = $this->auth->verification()->first(['token' , 'expires_at']);
+        if ($this->auth instanceof Admin){
+            $verification = $this->auth->verification()->first(['token' , 'expires_at']);
 
-        if($this->auth->hasVerifiedEmail())
-        {
-            return $this->errorResponse(message:'Email address is already verified.', statusCode: 400);
+            if($this->auth->hasVerifiedEmail())
+            {
+                return $this->errorResponse(message:'Email address is already verified.', statusCode: 400);
+            }
+
+            if (!$verification) {
+                return $this->errorResponse(message:'Verification record not found.', statusCode: 404);
+            }
+
+            if ($verification->token !== $token || now()->greaterThan($verification->expires_at)) {
+                return $this->errorResponse(message:'Invalid or expired token.', statusCode: 400);
+            }
+
+            $this->auth->markEmailAsVerified();
+
+            $this->auth->verification()->delete();
+
+            return $this->successResponse(message:'Email verified successfully.');
         }
+        return $this->errorResponse(message:'You are not authorized to perform this action.');
 
-        if (!$verification) {
-            return $this->errorResponse(message:'Verification record not found.', statusCode: 404);
-        }
-
-        if ($verification->token !== $token || now()->greaterThan($verification->expires_at)) {
-             return $this->errorResponse(message:'Invalid or expired token.', statusCode: 400);
-        }
-
-        $this->auth->markEmailAsVerified();
-
-        $this->auth->verification()->delete();
-
-        return $this->successResponse(message:'Email verified successfully.');
     }
 }
